Added new controller and replaced render logic into it
