+++ src/controllers/days-list.js
@@ -2,10 +2,9 @@ import NoEventsComponent from "../components/no-events.js";
 import TripInfoComponent from "../components/trip-info.js";
 import MenuComponent from "../components/menu.js";
 import FilterComponent from "../components/filter.js";
-import EventComponent from "../components/event.js";
 import SortComponent, {SortType} from "../components/sort.js";
+import PointController from "./point-controller.js";
 import DayComponent from "../components/day.js";
-import EventEditComponent from "../components/event-edit-form.js";
 import {render, replace, RenderPosition, remove} from "../utils/render.js";
 import {stringifyDate} from "../utils/util.js";
 
@@ -30,6 +29,7 @@ const getSortedEvents = (events, sortType) => {
   return sortedEvents;
 };
 
+
 const createDayComponents = (events, isSorting) => {
   if (isSorting) {
     return [new DayComponent(new Date(), true)];
@@ -53,43 +53,15 @@ const renderTripEvents = (events, container, isSorting = false) => {
     arr = arr.sort((a, b) => a.startTime - b.startTime);
   }
 
+  const pointController = new PointController(container);
+
   arr.forEach((event) => {
-    const eventComponent = new EventComponent(event);
-    const eventEditComponent = new EventEditComponent(`edit`, event);
-    const dayComponent = isSorting ? dayComponentsArray[0] : dayComponentsArray.find((day) => stringifyDate(day.date.startTime) === stringifyDate(event.startTime));
-
-    const eventToEditHandler = () => {
-      replace(eventEditComponent, eventComponent);
-    };
-
-    const editToEventHandler = () => {
-      replace(eventComponent, eventEditComponent);
-    };
-
-    const deleteEventHandler = () => {
-      remove(eventComponent);
-      remove(eventEditComponent);
-
-      if (dayComponent.getElement().querySelector(`ul`).children.length === 0) {
-        remove(dayComponent);
-      }
-    };
-
-    eventComponent.setOpenEditHandler(eventToEditHandler);
-    eventEditComponent.setEditFormHandlers(
-        editToEventHandler,
-        (evt) => {
-          evt.preventDefault();
-          editToEventHandler();
-        },
-        deleteEventHandler
-    );
-
-    render(dayComponent.getElement().querySelector(`ul`), eventComponent, RenderPosition.BEFOREEND);
+    pointController.render(event, dayComponentsArray, isSorting);
   });
 
   dayComponentsArray.sort((a, b) => a.date.startTime - b.date.startTime).forEach((dayComponent) => render(container, dayComponent, RenderPosition.BEFOREEND));
 };
+
 
 export default class TripController {
   constructor(containerComponent) {
